[modelica] Rename rules and fix rulesets
